add acl quotes acl.

+++ packages/Webkul/Admin/src/DataGrids/Activity/ActivityDataGrid.php
@@ -183,21 +183,25 @@ class ActivityDataGrid extends DataGrid
      */
     public function prepareActions(): void
     {
-        $this->addAction([
-            'index'  => 'edit',
-            'icon'   => 'icon-edit',
-            'title'  => trans('admin::app.activities.index.datagrid.edit'),
-            'method' => 'GET',
-            'url'    => fn ($row) => route('admin.activities.edit', $row->id),
-        ]);
+        if (bouncer()->hasPermission('activities.edit')) {
+            $this->addAction([
+                'index'  => 'edit',
+                'icon'   => 'icon-edit',
+                'title'  => trans('admin::app.activities.index.datagrid.edit'),
+                'method' => 'GET',
+                'url'    => fn ($row) => route('admin.activities.edit', $row->id),
+            ]);
+        }
 
-        $this->addAction([
-            'index'  => 'delete',
-            'icon'   => 'icon-delete',
-            'title'  => trans('admin::app.activities.index.datagrid.update'),
-            'method' => 'DELETE',
-            'url'    => fn ($row) => route('admin.activities.delete', $row->id),
-        ]);
+        if (bouncer()->hasPermission('activities.delete')) {
+            $this->addAction([
+                'index'  => 'delete',
+                'icon'   => 'icon-delete',
+                'title'  => trans('admin::app.activities.index.datagrid.update'),
+                'method' => 'DELETE',
+                'url'    => fn ($row) => route('admin.activities.delete', $row->id),
+            ]);
+        }
     }
 
     /**
